sifsearchbackend/requirements.txt
algoliasearch==3.0.0
algoliasearch-django==3.0.0
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.1
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.6
psycopg2-binary==2.9.9
pycparser==2.21
pycryptodome==3.20.0
PyJWT==2.8.0
Pyrebase4==4.7.1
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
urllib3==2.1.0